// web/src/api/types.gen.ts
export type GetApiProductsResponses = {
    /**
     * Default Response
     */
    200: {
        data: Array<{
            id: number;
            name: string;
            sku: string;
            category: 'electronics' | 'clothing' | 'food' | 'furniture' | 'tools' | 'other';
            price: number;
            quantity: number;
            minStock: number;
            status: 'in_stock' | 'low_stock' | 'out_of_stock';
            storeId: number;
            storeName: string;
            createdAt: string;
            updatedAt: string;
        }>;
        pagination: {
            page: number;
            limit: number;
            total: number;
            totalPages: number;
            hasNext: boolean;
            hasPrev: boolean;
        };
        links: {
            self: string;
            first: string;
            prev: string | null;
            next: string | null;
            last: string;
        };
    };
};

export type GetApiProductsResponse = GetApiProductsResponses[keyof GetApiProductsResponses];